+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-(none yet)
+- Added methods for retrieving task lists (lending requests and requested resources).
 
 ## [0.7.0] - 2018-09-02